feat: Cleaning utils & separating styles into /constants
